requirements.txt
kaggle
pandas 
numpy
ipykernel
psycopg2
sqlalchemy
python-dotenv